figued out OWNER.all and WALKER.all.
We had to add the string TYPE to our USERS table.

This commit saw creation of TWO migrations to remove the
UserTypes table and the type_id column from the users table.

The second migration added the string type to users table.

We updated our edited controllers and the seeds file to remove
all occurences of UserType